Embeddings update

+++ modules/database3.py
@@ -80,11 +80,24 @@ class SubscriptionManager:
             print(f"Subscription code {subscription_code} does not exist.")
         return False
 
-    def delete_chat_id(self, chat_id):
-        self.db.chat_ids.delete_one({
-            'chat_id': chat_id
+    def delete_chat_id(self, subscription_code, chat_id):
+        subscription = self.db.subscriptions.find_one({
+            'subscription_code': subscription_code
         })
 
+        if subscription:
+            try:
+                self.db.chat_ids.delete_one({
+                    'subscription_code': subscription_code,
+                    'chat_id': chat_id
+                })
+                return True
+            except Exception as e:
+                print(f"Error adding chat ID: {e}")
+        else:
+            print(f"Subscription code {subscription_code} does not exist.")
+        return False
+    
     def verify_subscription_code(self, subscription_code):
         result = self.db.subscriptions.find_one({
             'subscription_code': subscription_code
